refactor(audit): split paths and proof god-files (#5645)
