+++ src/core/paths.rs
@@ -4,9 +4,13 @@ use std::ffi::OsString;
 use std::path::{Component, Path, PathBuf};
 use std::sync::{Mutex, OnceLock};
 
+mod locations;
 mod rigs;
+mod runtime;
 
+pub use locations::*;
 pub use rigs::*;
+pub use runtime::*;
 
 fn artifact_root_override() -> &'static Mutex<Option<PathBuf>> {
     static OVERRIDE: OnceLock<Mutex<Option<PathBuf>>> = OnceLock::new();
@@ -240,98 +244,6 @@ pub fn resolve_contained_local_path(
             ]),
         ))
     }
-}
-
-/// Projects directory
-pub fn projects() -> Result<PathBuf> {
-    Ok(homeboy()?.join("projects"))
-}
-
-/// Project directory path (e.g., ~/.config/homeboy/projects/{id}/)
-pub fn project_dir(id: &str) -> Result<PathBuf> {
-    Ok(projects()?.join(id))
-}
-
-/// Project config file path (e.g., ~/.config/homeboy/projects/{id}/{id}.json)
-pub fn project_config(id: &str) -> Result<PathBuf> {
-    Ok(projects()?.join(id).join(format!("{}.json", id)))
-}
-
-/// Servers directory
-pub fn servers() -> Result<PathBuf> {
-    Ok(homeboy()?.join("servers"))
-}
-
-/// Components directory
-pub fn components() -> Result<PathBuf> {
-    Ok(homeboy()?.join("components"))
-}
-
-/// Extensions directory
-pub fn extensions() -> Result<PathBuf> {
-    Ok(homeboy()?.join("extensions"))
-}
-
-/// Agent runtime package directory
-pub fn agent_runtimes() -> Result<PathBuf> {
-    Ok(homeboy()?.join("agent-runtimes"))
-}
-
-/// Keys directory
-pub fn keys() -> Result<PathBuf> {
-    Ok(homeboy()?.join("keys"))
-}
-
-/// Backups directory
-pub fn backups() -> Result<PathBuf> {
-    Ok(homeboy()?.join("backups"))
-}
-
-/// Daemon runtime state directory (~/.config/homeboy/daemon/).
-fn daemon_state_dir() -> Result<PathBuf> {
-    Ok(homeboy()?.join("daemon"))
-}
-
-/// Daemon runtime state file (~/.config/homeboy/daemon/state.json).
-pub fn daemon_state_file() -> Result<PathBuf> {
-    Ok(daemon_state_dir()?.join("state.json"))
-}
-
-/// Daemon durable job state file (~/.config/homeboy/daemon/jobs.json).
-pub fn daemon_jobs_file() -> Result<PathBuf> {
-    Ok(daemon_state_dir()?.join("jobs.json"))
-}
-
-/// Runner connection session state directory (~/.config/homeboy/runner-sessions/).
-fn runner_sessions_dir() -> Result<PathBuf> {
-    Ok(homeboy()?.join("runner-sessions"))
-}
-
-/// Runner connection session state file (~/.config/homeboy/runner-sessions/{id}.json).
-pub fn runner_session_file(id: &str) -> Result<PathBuf> {
-    Ok(runner_sessions_dir()?.join(format!("{}.json", id)))
-}
-
-/// Managed service tunnel runtime state directory (~/.local/share/homeboy/service-tunnels/{id}/).
-pub fn service_tunnel_runtime_dir(id: &str) -> Result<PathBuf> {
-    Ok(homeboy_data()?
-        .join("service-tunnels")
-        .join(sanitize_path_segment(id)))
-}
-
-/// Managed service tunnel runtime state file.
-pub fn service_tunnel_runtime_state_file(id: &str) -> Result<PathBuf> {
-    Ok(service_tunnel_runtime_dir(id)?.join("state.json"))
-}
-
-/// Preview ingress route declarations (~/.config/homeboy/preview-ingress/routes/).
-pub fn preview_ingress_routes_dir() -> Result<PathBuf> {
-    Ok(homeboy()?.join("preview-ingress").join("routes"))
-}
-
-/// Preview ingress route declaration file.
-pub fn preview_ingress_route_file(id: &str) -> Result<PathBuf> {
-    Ok(preview_ingress_routes_dir()?.join(format!("{}.json", sanitize_path_segment(id))))
 }
 
 /// Extension directory path
